backend/requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.0
aiofiles>=23.2.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Image Processing
opencv-python-headless>=4.8.0
numpy>=1.26.0
Pillow>=10.1.0

# HTTP Client
httpx>=0.25.0

# Environment
python-dotenv>=1.0.0
